fix: datatype Error

+++ src/main/java/com/example/trainee_app/controller/StudentController.java
@@ -29,20 +29,20 @@ public class StudentController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Student> getById(@PathVariable Long id) {
+    public ResponseEntity<Student> getById(@PathVariable int id) {
         return studentService.getStudentById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Student> update(@PathVariable Long id,
+    public ResponseEntity<Student> update(@PathVariable int id,
                                           @Valid @RequestBody Student student) {
         return ResponseEntity.ok(studentService.updateStudent(id, student));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> delete(@PathVariable Long id) {
+    public ResponseEntity<String> delete(@PathVariable int id) {
         studentService.deleteStudent(id);
         return ResponseEntity.ok("Student deleted successfully.");
     }
